fix: d3d12 agility and other various fixes

+++ src/DX12/DX12ResourceLayout.cpp
@@ -35,8 +35,10 @@ u32 DX12ResourceLayout::GetMaxLocalConstantSize() const
     // Each global constant descriptor takes up 2 DWORDs in the root signature (as root descriptor).
     // There is the option of using a descriptor table for constants to reduce their size, but adding a level of
     // indirection, this is probably not needed thanks to bindless existing nowadays!
-    return std::max<u32>(0, (featureChecker.GetMaxRootSignatureDWORDSize() - 2 * globalConstants.size())) *
-           sizeof(DWORD);
+    return std::max<u32>(
+               0,
+               (featureChecker.GetMaxRootSignatureDWORDSize() - 2 * static_cast<u32>(globalConstants.size()))) *
+           static_cast<u32>(sizeof(DWORD));
 }
 
 ComPtr<ID3D12RootSignature>& DX12ResourceLayout::GetRootSignature()
